+++ packages/subagents/CHANGELOG.md
@@ -11,6 +11,9 @@
 
 ## [Unreleased]
 
+### Fixed
+- Fixed builtin subagent skill resolution from project cwd ([#1087](https://github.com/flora131/atomic/issues/1087)).
+
 ## [0.8.18] - 2026-05-27
 
 ### Changed